[FEAT] Fix buy workflow PR comments
