+++ imx-android-sdk/src/main/java/com/immutable/sdk/utils/TokenType.kt
@@ -1,5 +1,5 @@
 package com.immutable.sdk.utils
 
-enum class TokenType {
+internal enum class TokenType {
     ETH, ERC20, ERC721, MINTABLE_ERC20, MINTABLE_ERC721
 }
